Implement Eval trait
- Allow multiple different evaluation functions
