+++ eval/src/lib.rs
@@ -1,1 +1,6 @@
+pub use crate::eval::{
+    Eval, Score, CHECKMATE_BLACK, CHECKMATE_WHITE, EQUAL_POSITION, NEGATIVE_INF, POSITIVE_INF,
+};
+
 pub mod eval;
+pub mod material_mobility;
