+++ search/src/alpha_beta_entry.rs
@@ -1,5 +1,5 @@
 use crate::search::MAX_SEARCH_DEPTH;
-use eval::eval::Score;
+use eval::Score;
 use movegen::r#move::Move;
 use movegen::zobrist::Zobrist;
 use std::mem;
