+++ search/src/negamax.rs
@@ -2,8 +2,9 @@ use crate::negamax_entry::NegamaxEntry;
 use crate::search::{Search, SearchCommand, SearchInfo, SearchResult, REPETITIONS_TO_DRAW};
 use crate::search_data::SearchData;
 use crossbeam_channel::{Receiver, Sender};
-use eval::eval::{Eval, CHECKMATE_BLACK, CHECKMATE_WHITE, EQUAL_POSITION, NEGATIVE_INF};
+use eval::{Score, CHECKMATE_BLACK, CHECKMATE_WHITE, EQUAL_POSITION, NEGATIVE_INF};
 use movegen::move_generator::MoveGenerator;
+use movegen::position::Position;
 use movegen::position_history::PositionHistory;
 use movegen::r#move::{Move, MoveList};
 use movegen::side::Side;
@@ -14,6 +15,7 @@ use std::time::Instant;
 type NegamaxTable = TranspositionTable<Zobrist, NegamaxEntry>;
 
 pub struct Negamax {
+    eval_relative: fn(&Position) -> Score,
     transpos_table: NegamaxTable,
 }
 
@@ -76,9 +78,10 @@ impl Search for Negamax {
 }
 
 impl Negamax {
-    pub fn new(table_idx_bits: usize) -> Self {
+    pub fn new(eval_relative: fn(&Position) -> Score, table_idx_bits: usize) -> Self {
         assert!(table_idx_bits > 0);
         Self {
+            eval_relative,
             transpos_table: NegamaxTable::new(table_idx_bits),
         }
     }
@@ -186,7 +189,7 @@ impl Negamax {
 
         search_data.increment_eval_calls();
         let pos = search_data.pos_history().current_pos();
-        let mut score = Eval::eval_relative(pos);
+        let mut score = (self.eval_relative)(pos);
         let mut best_score = score;
         let mut best_move = Move::NULL;
