+++ search/tests/search.rs
@@ -1,5 +1,6 @@
 use crossbeam_channel::{unbounded, Receiver};
-use eval::eval::{Eval, CHECKMATE_BLACK, CHECKMATE_WHITE, EQUAL_POSITION, NEGATIVE_INF};
+use eval::material_mobility::MaterialMobility;
+use eval::{Eval, Score, CHECKMATE_BLACK, CHECKMATE_WHITE, EQUAL_POSITION, NEGATIVE_INF};
 use movegen::fen::Fen;
 use movegen::move_generator::MoveGenerator;
 use movegen::piece;
@@ -15,6 +16,8 @@ use search::searcher::Searcher;
 use std::cmp;
 use std::time::Duration;
 
+const EVAL_ABSOLUTE: fn(pos: &Position) -> Score = MaterialMobility::eval;
+const EVAL_RELATIVE: fn(pos: &Position) -> Score = MaterialMobility::eval_relative;
 const TABLE_IDX_BITS: usize = 20;
 const TIMEOUT_PER_TEST: Duration = Duration::from_millis(30000);
 
@@ -179,7 +182,7 @@ fn search_quiescence(search_algo: impl Search + Send + 'static) {
     let mut max_score = NEGATIVE_INF;
     for m in move_list.iter() {
         pos_history.do_move(*m);
-        max_score = cmp::max(max_score, Eval::eval(pos_history.current_pos()));
+        max_score = cmp::max(max_score, EVAL_ABSOLUTE(pos_history.current_pos()));
         pos_history.undo_last_move();
     }
     assert_eq!(max_score, tester.search(pos_history.clone(), depth).score());
@@ -327,57 +330,60 @@ fn negamax_search_results_independent_of_transposition_table_size() {
     let max_depth = 2;
     let mut searchers = Vec::new();
     for table_idx_bits in 1..=max_table_idx_bits {
-        searchers.push(SearchTester::new(Negamax::new(table_idx_bits)));
+        searchers.push(SearchTester::new(Negamax::new(
+            EVAL_RELATIVE,
+            table_idx_bits,
+        )));
     }
     search_results_equal(min_depth, max_depth, searchers);
 }
 
 #[test]
 fn negamax_checkmate_white() {
-    let negamax = Negamax::new(TABLE_IDX_BITS);
+    let negamax = Negamax::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     checkmate_white(negamax);
 }
 
 #[test]
 fn negamax_checkmate_black() {
-    let negamax = Negamax::new(TABLE_IDX_BITS);
+    let negamax = Negamax::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     checkmate_black(negamax);
 }
 
 #[test]
 fn negamax_stalemate() {
-    let negamax = Negamax::new(TABLE_IDX_BITS);
+    let negamax = Negamax::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     stalemate(negamax);
 }
 
 #[test]
 fn negamax_search_quiescence() {
-    let negamax = Negamax::new(TABLE_IDX_BITS);
+    let negamax = Negamax::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     search_quiescence(negamax);
 }
 
 #[test]
 fn negamax_pv_valid_after_hash_table_hit_depth_1() {
-    let negamax = Negamax::new(TABLE_IDX_BITS);
+    let negamax = Negamax::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     pv_valid_after_hash_table_hit_depth_1(negamax);
 }
 
 #[test]
 #[ignore]
 fn negamax_pv_valid_after_hash_table_hit_depth_greater_than_1() {
-    let negamax = Negamax::new(TABLE_IDX_BITS);
+    let negamax = Negamax::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     pv_valid_after_hash_table_hit_depth_greater_than_1(negamax);
 }
 
 #[test]
 fn negamax_play_threefold_repetition_in_losing_position() {
-    let negamax = Negamax::new(TABLE_IDX_BITS);
+    let negamax = Negamax::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     play_threefold_repetition_in_losing_position(negamax);
 }
 
 #[test]
 fn negamax_avoid_threefold_repetition_in_winning_position() {
-    let negamax = Negamax::new(TABLE_IDX_BITS);
+    let negamax = Negamax::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     avoid_threefold_repetition_in_winning_position(negamax);
 }
 
@@ -393,45 +399,48 @@ fn alpha_beta_search_results_independent_of_transposition_table_size() {
     let max_depth = 5;
     let mut searchers = Vec::new();
     for table_idx_bits in 1..=max_table_idx_bits {
-        searchers.push(SearchTester::new(AlphaBeta::new(table_idx_bits)));
+        searchers.push(SearchTester::new(AlphaBeta::new(
+            EVAL_RELATIVE,
+            table_idx_bits,
+        )));
     }
     search_results_equal(min_depth, max_depth, searchers);
 }
 
 #[test]
 fn alpha_beta_checkmate_white() {
-    let alpha_beta = AlphaBeta::new(TABLE_IDX_BITS);
+    let alpha_beta = AlphaBeta::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     checkmate_white(alpha_beta);
 }
 
 #[test]
 fn alpha_beta_checkmate_black() {
-    let alpha_beta = AlphaBeta::new(TABLE_IDX_BITS);
+    let alpha_beta = AlphaBeta::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     checkmate_black(alpha_beta);
 }
 
 #[test]
 fn alpha_beta_stalemate() {
-    let alpha_beta = AlphaBeta::new(TABLE_IDX_BITS);
+    let alpha_beta = AlphaBeta::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     stalemate(alpha_beta);
 }
 
 #[test]
 fn alpha_beta_search_quiescence() {
-    let alpha_beta = AlphaBeta::new(TABLE_IDX_BITS);
+    let alpha_beta = AlphaBeta::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     search_quiescence(alpha_beta);
 }
 
 #[test]
 fn alpha_beta_pv_valid_after_hash_table_hit_depth_1() {
-    let alpha_beta = AlphaBeta::new(TABLE_IDX_BITS);
+    let alpha_beta = AlphaBeta::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     pv_valid_after_hash_table_hit_depth_1(alpha_beta);
 }
 
 #[test]
 #[ignore]
 fn alpha_beta_pv_valid_after_hash_table_hit_depth_greater_than_1() {
-    let alpha_beta = AlphaBeta::new(TABLE_IDX_BITS);
+    let alpha_beta = AlphaBeta::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     pv_valid_after_hash_table_hit_depth_greater_than_1(alpha_beta);
 }
 
@@ -441,18 +450,18 @@ fn alpha_beta_count_searched_nodes_middlegame_position() {
     let fen = "r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R w KQkq - 0 1";
     let pos = Fen::str_to_pos(fen).unwrap();
     let depth = 6;
-    let alpha_beta = AlphaBeta::new(TABLE_IDX_BITS);
+    let alpha_beta = AlphaBeta::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     count_searched_nodes(alpha_beta, pos, depth);
 }
 
 #[test]
 fn alpha_beta_play_threefold_repetition_in_losing_position() {
-    let alpha_beta = AlphaBeta::new(TABLE_IDX_BITS);
+    let alpha_beta = AlphaBeta::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     play_threefold_repetition_in_losing_position(alpha_beta);
 }
 
 #[test]
 fn alpha_beta_avoid_threefold_repetition_in_winning_position() {
-    let alpha_beta = AlphaBeta::new(TABLE_IDX_BITS);
+    let alpha_beta = AlphaBeta::new(EVAL_RELATIVE, TABLE_IDX_BITS);
     avoid_threefold_repetition_in_winning_position(alpha_beta);
 }
